#249 - Restructure services folders and files, fix imports
