+++ src/services/tests/change-requests.api.test.ts
@@ -5,11 +5,11 @@
 
 import { rest } from 'msw';
 import { setupServer } from 'msw/node';
-import { apiUrls } from '../../../shared/urls';
+import { apiUrls } from '../../shared/urls';
 import {
   exampleAllChangeRequests,
   exampleStageGateChangeRequest
-} from '../../../test-support/test-data/change-requests.stub';
+} from '../../test-support/test-data/change-requests.stub';
 import { getAllChangeRequests, getSingleChangeRequest } from '../change-requests.api';
 
 // Mock the server endpoint(s) that the component will hit
