+++ src/services/tests/change-requests.hooks.test.tsx
@@ -6,16 +6,16 @@
 import { renderHook } from '@testing-library/react-hooks';
 import { AxiosResponse } from 'axios';
 import { ChangeRequest } from 'utils';
-import { queryClientProviderWrapper as wrapper } from '../../../test-support/test-utils';
-import { mockPromiseAxiosResponse } from '../../../test-support/test-data/test-utils.stub';
+import { queryClientProviderWrapper as wrapper } from '../../test-support/test-utils';
+import { mockPromiseAxiosResponse } from '../../test-support/test-data/test-utils.stub';
 import {
   exampleAllChangeRequests,
   exampleStageGateChangeRequest
-} from '../../../test-support/test-data/change-requests.stub';
-import { getAllChangeRequests, getSingleChangeRequest } from '../../apis/change-requests.api';
+} from '../../test-support/test-data/change-requests.stub';
+import { getAllChangeRequests, getSingleChangeRequest } from '../change-requests.api';
 import { useAllChangeRequests, useSingleChangeRequest } from '../change-requests.hooks';
 
-jest.mock('../../apis/change-requests.api');
+jest.mock('../change-requests.api');
 
 describe('change request hooks', () => {
   it('handles getting a list of change requests', async () => {
