+++ src/services/tests/projects.hooks.test.tsx
@@ -6,14 +6,14 @@
 import { renderHook } from '@testing-library/react-hooks';
 import { AxiosResponse } from 'axios';
 import { Project } from 'utils';
-import { queryClientProviderWrapper as wrapper } from '../../../test-support/test-utils';
-import { mockPromiseAxiosResponse } from '../../../test-support/test-data/test-utils.stub';
-import { exampleAllProjects, exampleProject1 } from '../../../test-support/test-data/projects.stub';
-import { exampleWbsProject1 } from '../../../test-support/test-data/wbs-numbers.stub';
-import { getAllProjects, getSingleProject } from '../../apis/projects.api';
+import { queryClientProviderWrapper as wrapper } from '../../test-support/test-utils';
+import { mockPromiseAxiosResponse } from '../../test-support/test-data/test-utils.stub';
+import { exampleAllProjects, exampleProject1 } from '../../test-support/test-data/projects.stub';
+import { exampleWbsProject1 } from '../../test-support/test-data/wbs-numbers.stub';
+import { getAllProjects, getSingleProject } from '../projects.api';
 import { useAllProjects, useSingleProject } from '../projects.hooks';
 
-jest.mock('../../apis/projects.api');
+jest.mock('../projects.api');
 
 describe('project hooks', () => {
   it('handles getting a list of projects', async () => {
