+++ src/services/tests/users.hooks.test.tsx
@@ -6,13 +6,13 @@
 import { renderHook } from '@testing-library/react-hooks';
 import { AxiosResponse } from 'axios';
 import { User } from 'utils';
-import { queryClientProviderWrapper as wrapper } from '../../../test-support/test-utils';
-import { mockPromiseAxiosResponse } from '../../../test-support/test-data/test-utils.stub';
-import { exampleAllUsers, exampleAdminUser } from '../../../test-support/test-data/users.stub';
-import { getAllUsers, getSingleUser } from '../../apis/users.api';
+import { queryClientProviderWrapper as wrapper } from '../../test-support/test-utils';
+import { mockPromiseAxiosResponse } from '../../test-support/test-data/test-utils.stub';
+import { exampleAllUsers, exampleAdminUser } from '../../test-support/test-data/users.stub';
+import { getAllUsers, getSingleUser } from '../users.api';
 import { useAllUsers, useSingleUser } from '../users.hooks';
 
-jest.mock('../../apis/users.api');
+jest.mock('../users.api');
 
 describe('user hooks', () => {
   it('handles getting a list of users', async () => {
