+++ src/services/tests/work-packages.api.test.ts
@@ -5,11 +5,11 @@
 
 import { rest } from 'msw';
 import { setupServer } from 'msw/node';
-import { apiUrls } from '../../../shared/urls';
+import { apiUrls } from '../../shared/urls';
 import {
   exampleAllWorkPackages,
   exampleWorkPackage1
-} from '../../../test-support/test-data/work-packages.stub';
+} from '../../test-support/test-data/work-packages.stub';
 import { getAllWorkPackages, getSingleWorkPackage } from '../work-packages.api';
 
 // Mock the server endpoint(s) that the component will hit
